+++ src/components/modal/AddCardModal.tsx
@@ -1,0 +1,48 @@
+const AddCardModal = () => {
+  return (
+    <div className="root">
+      <div className="modal-dimmed">
+        <div className="modal">
+          <div className="flex-center">
+            <div className="modal-item-container">
+              <div className="modal-item-dot"></div>
+              <span className="modal-item-name">클린 카드</span>
+            </div>
+            <div className="modal-item-container">
+              <div className="modal-item-dot"></div>
+              <span className="modal-item-name">클린 카드</span>
+            </div>
+            <div className="modal-item-container">
+              <div className="modal-item-dot"></div>
+              <span className="modal-item-name">클린 카드</span>
+            </div>
+            <div className="modal-item-container">
+              <div className="modal-item-dot"></div>
+              <span className="modal-item-name">클린 카드</span>
+            </div>
+          </div>
+          <div className="flex-center">
+            <div className="modal-item-container">
+              <div className="modal-item-dot"></div>
+              <span className="modal-item-name">클린 카드</span>
+            </div>
+            <div className="modal-item-container">
+              <div className="modal-item-dot"></div>
+              <span className="modal-item-name">클린 카드</span>
+            </div>
+            <div className="modal-item-container">
+              <div className="modal-item-dot"></div>
+              <span className="modal-item-name">클린 카드</span>
+            </div>
+            <div className="modal-item-container">
+              <div className="modal-item-dot"></div>
+              <span className="modal-item-name">클린 카드</span>
+            </div>
+          </div>
+        </div>
+      </div>
+    </div>
+  );
+};
+
+export default AddCardModal;
