+++ src/pages/confirmPage/AfterCardPage.tsx
